Refactor Command Router based command consumer.
The functionality of the ProtonBasedCommandRouterCommandConsumerFactoryImpl
class has been moved into separate classes. One generic
CommandRouterCommandConsumerFactory and a proton based consumer class,
providing the receiver on the internal command endpoint targeted by the
Command Router. This is a preparation for adding a corresponding
Kafka based consumer.
